@@ -20,25 +20,26 @@
         }
     },
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.1",
         "ext-zlib": "*",
         "ext-json": "*",
         "aws/aws-sdk-php": "^3.209",
         "defuse/php-encryption": "^2.3",
-        "keboola/azure-key-vault-client": "^2.0|^3.0",
+        "keboola/azure-key-vault-client": "^3.0",
         "keboola/common-exceptions": "^1.2",
-        "vkartaviy/retry": "^0.2.2"
+        "vkartaviy/retry": "^0.2"
     },
     "require-dev": {
-        "infection/infection": "^0.18.2",
+        "infection/infection": "^0.26",
         "keboola/coding-standard": "^14.0",
         "phpstan/phpstan": "^1.8",
-        "phpstan/phpstan-phpunit": "^1.0.0",
+        "phpstan/phpstan-phpunit": "^1.0",
         "phpunit/phpunit": "^9.5",
         "sempro/phpunit-pretty-print": "^1.4",
-        "symfony/dotenv": "^5.4"
+        "symfony/dotenv": "^5.4|^6.0"
     },
     "scripts": {
+        "pre-autoload-dump": "Aws\\Script\\Composer\\Composer::removeUnusedServices",
         "tests": "phpunit --coverage-clover /tmp/build-logs/clover.xml --coverage-xml=/tmp/build-logs/coverage-xml --log-junit=/tmp/build-logs/phpunit.junit.xml",
         "phpcs": "phpcs --extensions=php src tests",
         "phpstan": "phpstan analyse --no-progress -c phpstan.neon",
@@ -61,5 +62,12 @@
             "infection/extension-installer": true,
             "dealerdirect/phpcodesniffer-composer-installer": true
         }
+    },
+    "extra": {
+        "aws/aws-sdk-php": [
+            "S3",
+            "Kms",
+            "Sts"
+        ]
     }
 }
